Move `Metric` into it's own file.
